ogl: rename trackball module into mathutils as it better reflects its purpose
